+++ mmvb_backend/toy_ais/implementations/uniform_random_conditions.py
@@ -27,30 +27,3 @@ class UniformRandomConditions(ToyAI):
         else:
             response.update(result)
         return response
-
-
-class WeightedRandomConditions(ToyAI):
-    @classproperty
-    def name(cls) -> str:
-        return "Weighted Random Conditions Solver"
-
-    @classproperty
-    def slug_name(cls) -> str:
-        return "weighted-random-conditions"
-
-    @classmethod
-    def health_check(cls, *args, **kwargs):
-        return {"data": f"OK from {cls.name}"}
-
-    @classmethod
-    def solve_case(cls, payload, *args, **kwargs):
-        response = {"error": "", "triage": None, "conditions": []}
-        try:
-            result = solve_case_random_conditions(
-                payload, "probability_weighted"
-            )
-        except Exception as exc:
-            response["error"] = f"Error handling case data. Got {str(exc)}"
-        else:
-            response.update(result)
-        return response
