+++ requirements.txt
@@ -16,4 +16,3 @@ redis==3.4.1
 stringcase==1.2.0
 pre-commit==2.2.0
 numpy==1.16.4
-requests==2.23.0
